ml_api/requirements.txt
flask
opencv-python
ultralytics
numpy
mediapipe
python-dotenv
dlib
Pillow
